feat: update auth routes
