+++ components/AuthForm/SignInForm.tsx
@@ -41,7 +41,7 @@ export const SignInForm = () => {
       // and redirect the user
       if (signInAttempt.status === "complete") {
         await setActive({ session: signInAttempt.createdSessionId });
-        router.push("/");
+        router.push("/companies");
       } else {
         // If the status is not complete, check why. User may need to
         // complete further steps.
