+++ components/AuthForm/index.tsx
@@ -4,12 +4,16 @@ import { Card, CardBody } from "@nextui-org/card";
 import { SignInForm } from "./SignInForm";
 import { SignUpForm } from "./SignUpForm";
 
-export const AuthForm = () => {
+type AuthFormProps = {
+  activeTab?: "sign-in" | "sign-up";
+};
+
+export const AuthForm = ({ activeTab = "sign-in" }: AuthFormProps) => {
   return (
     <div className="flex items-center flex-col">
       <Card className="max-w-sm w-full">
         <CardBody className="flex items-center">
-          <Tabs aria-label="Options">
+          <Tabs aria-label="Options" defaultSelectedKey={activeTab}>
             <Tab key="sign-in" className="w-full text-center" title="Sign in">
               <h1 className="text-2xl font-bold mt-4">Welcome back</h1>
               <SignInForm />
